Restructured /client/src folder
fix constants
