+++ client/src/AddFriend/Settings.js
